добавляет модули
