+++ index.html
@@ -234,7 +234,7 @@
       </section>
     </template>
 
-  <script src="js/functions.js"></script>
+  <script type="module" src="js/main.js"></script>
 
 </body>
 </html>
